pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps/", "config/"]

[project]
name = "pybot"
version = "0.0.1"
description = "A bot to manage your association server."
readme = "README.md"
license = "MIT"
keywords = ["discord", "bot", "association", "server"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat",
    "Topic :: Internet",
    "Topic :: Utilities"
]
dependencies = [
    "discord.py",
    "django",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-vv --tb=short --cov=apps --cov-report=term-missing"
DJANGO_SETTINGS_MODULE = "config.settings"
filterwarnings = "ignore::DeprecationWarning"

[tool.coverage.run]
omit = ["**/tests/*", "**/__init__.py"]
